[docs] Optimize images (mostly losslessly)
Several of the images in the docs were unoptimized. We can optimize them to reduce file size to slightly speed up the docs and also to reduce the size of the repo if we were to rewrite the history, though that is secondary.

Ran `optipng -o7` on every PNG file in the docs that is in our own repo (not in a submodule). One of the PNG files was a JPEG file so I renamed it to .jpg and optimized it with no noticeable loss of quality.

On JPEGs I ran `cjpeg -q 75` (note: mozjpeg's cjpeg) on most files and visually compared them. I also looked at https://tinyjpg.com, which is quite effective.

Total size reduction is 8.54 MiB. Not game-changing, but these were mostly lossless reductions. (There are bigger savings with GIFs and videos.)

Test plan: run the docs and browse around through the EAS Build guides.